poprawka do rejestracji + historia + wyprzedaże

+++ src/main/java/com/falynsky/smartmarkt/utils/PriceUtils.java
@@ -3,6 +3,6 @@ package com.falynsky.smartmarkt.utils;
 public class PriceUtils {
 
     public static double roundPrice(double price) {
-        return (double) (Math.round(price * 100.0) / 100.0);
+        return Math.round(price * 100.0) / 100.0;
     }
 }
